requirements.txt
django-role-permissions
-r requirements-dev.txt
gunicorn
psycopg2